Bumped version number to 3.0.9 Fixed up dist release 32 bit config. build.rb now builds 32 bit config as well.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@2227 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/License.cpp
@@ -83,7 +83,7 @@ const std::string License::decodeBase64(const std::string& data_)
 {
 	const std::string data = ensureNewLinesPresent(data_);
 
-	if(data.size() > std::numeric_limits<int>::max())
+	if(data.size() > (size_t)std::numeric_limits<int>::max())
 		throw License::LicenseExcep("Data too big");
 
     try {
